ProjectExplorer: compress rebuild contents of TargetGroupItem
An updated toolchain can affect multiple kits and for each kit the
content of the TargetGroupItem got rebuilt which in turn completely
rebuilds the project panel. Taking up to 5 seconds in the main thread on
some machines after loading a session. Do the Rebuild only once per
eventloop roundtrip reduces this to 300ms.

Change-Id: I2c2006b280c8d046ac79e8f016e2080eb0796796
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>