ProjectExplorer: fix resetting gcc install dir
If m_installDir was not empty when GccToolchain::resetToolchain gets
called the install dir was not reset since GccToolchain::installDir
would just return m_installDir in that case. But since the install dir
is not needed in GccToolchain::resetToolchain we can just clear the
install dir and rely on the fact that it gets autodected when
GccToolchain::installDir gets called.

This reduces the startup time of Qt Creator by around 700 ms on some
machines since the android toolchain code relies on calling
GccToolchain::resetToolchain in order to pick up JDK changes.

Change-Id: Ic69bca3946a91388b793317050a0b53dbf0cea2d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>