EffectComposer: Use text values for adding properties directly
On macOS value editing signal is not emitted if the user directly pushes
the apply button. (The editing field stays on editing mode, so the
changes are not committed)
Also if the name is not edited, and display name is not changed, if the
user pushes apply, the uniName is edited on windows, but is not
displayed while the semi-dialog is opened,  which is not good.

Fixes: QDS-14373
Change-Id: I5b1ac0b906baf775611564fb5021103c29e03f73
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>