MesonProjectManager: Rework project tree
This commit address the following issues:
- when using QTC and Meson from containers user home directory tend
to be mixed up with both /home/... and /var/home/..., using
canonicalPath removes all simlinks.
- source files are now under target node which makes more obvious
which sources are used by which target. Also before, N targets
in the same directory using the same source file makes it appear
N times in the same node.
- decoupling files sorting and nodes building makes simpler to
implement several tree layout strategies.

When sorting files per target, non source files that are not direct
children of the target folder are displayed outside of the target folder.
This allows to display files like README that are declared as target
extra_files directly where they are located in the file tree.

Change-Id: I543ccc311b08b0f360904a693661cc378e42debe
Reviewed-by: hjk <hjk@qt.io>