Terminal: Fix resize logic
When a terminal window is resized (e.g. on first show) it needs
to communicate its size to the PTY process. If the process is not
yet running we need to make sure that the size is set again once
it is running.

Adding a bool return value to "resizePty" allows us to detect if
the size change was applied, which allows us to update the
surface size only if the pty was also changed.

With this we can use the "liveSize" of the surface to check
if a resizeEvent needs to be passed on to the PTY, which will
be triggered by the Process::started signal once the process
is started.

Fixes: QTCREATORBUG-32290
Change-Id: I613275e75d343ccb357c2a797d096f0e1f96fed7
Reviewed-by: Cristian Adam <cristian.adam@qt.io>