Terminal: Fix freeze when closing terminal on newer Windows 11
Windows 11 24H2 changed the underlying implementation and the
"ClosePseudoConsole" code that we copied from the Terminal application
didn't work as it did before.

The new Terminal releases also removes the WaitForSingleObject call.
See https://github.com/microsoft/terminal/pull/17704

We cannot rely on the ConPty API since it doesn't send the "--
resizeQuirk" parameter to "conhost.exe", and the QTCREATORBUG-30007
would be resurrected.

Fixes: QTCREATORBUG-32192
Change-Id: Ie21833a1eb21acd0ba21109f3dbc28595b0d94a4
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>