CMake: Fix macOS rpath removal when building against non-frameworks
When Qt Creator is built against a non-framework macOS Qt, the build
system attempts to remove build rpaths from the installed binary
rpaths.

In case of shared library Qt dependencies (non-framework) the code
attempts to remove the whole absolute path of the Qt library as an
rpath. rpaths are usually not absolute file paths though, but rather
directory paths. This results in the removal of the wrong rpath,
causing an error at install time.

Instead of removing the whole absolute path, remove only the
directory part of library path. This is similar to what is done for
frameworks.

Amends 98db9774f2363314f51b1732646396112418320e

Task-number: QTCREATORBUG-24021
Change-Id: Ic247243a780bacfe7ba30f3933cdcd07de4a531c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>