Terminal: Fix Id
m_commandLine is not useful to distinguish terminals
as it may contain wrapper scripts with temporary arguments that make each invocation of a target
spawn a new terminal window.

Instead use the provided "shell name".

Also make sure to use the actual target name as the
shell name in runcontrol.

Fixes: QTCREATORBUG-32197
Change-Id: I8c0760be375d559391dbd831080aa7153786f411
Reviewed-by: Cristian Adam <cristian.adam@qt.io>