ClangFormat et al.: Refactor settings factories
Excessive granularity in style settings factories caused issues
when individual widgets of the settings page were interdependent.
For example, in ClangFormat settings the exporter widget
(previously created independently) must know the currently selected mode
from the global settings widget (previously also created independently)
in order to export the correct settings
(either legacy indenter's XML configuration or clang-format's YAML).
This change shifts the granularity by one level
(onto `CodeStyleEditor` class) so that the factories now produce
ready-to-use widgets wheres each of `CodeStyleEditor` subclasses
can handle dependencies in its building blocks as needed.

This patch also features some clean-ups in various places in ClangFormat
plugin as well as in some related parts of other plugins.
A couple of the most prominent examples are:
  * ClangFormat settings widget and factory are no longer subclasses
    of the built-in indenter's counterparts. Instead the ClangFormat
    settings widget now *contains* the built-in indenter settings
    as a component.
  * Consequently, ClangFormat won't parasitize the built-in
    indenter's settings by injecting its own settings tab into them
    and showing/hiding it or the others.
    See the occurrences of "additionalTab" throughout the code.

Note that the suggestion from QTCREATORBUG-32110 to add .clang-format
file name when exporting settings does not work with the native file
save dialog on Windows - the name is empty in the dialog
(I assume the reason is that Windows interprets it as a file extension
and automatically hides it). The non-native dialog is fine. It also
works more or less as expected when exporting the built-in indenter
XML configuration files.

Task-number: QTCREATORBUG-32110
Task-number: QTCREATORBUG-32112
Task-number: QTCREATORBUG-32113
Change-Id: I6651d81ba5e422692b58661e1e216f98923e70e0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>