Kits: Prevent long names from overly growing preferences
The combo boxes for device names, toolchains, debuggers and others were
setting their size to the length of the content. That meant that long
names for these items lead to a huge horizontal size for the preference
page content, leading to a lot of horizontal scrolling.

Set the default for these combo boxes to something smaller. They still
grow when resizing the preferences dialog.

Change-Id: I95869b3ccdac049bdd1f7100af648e5ac5932745
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>