Utils: Add a way to (shutdown-)guard pre-created objects
Some classes use private c'tors and a friend function for creation.

Since friendship is not transitive this cannot (nicely) be extended
to the GuardedObject implementation.

In the quest to achieve a uniform 'void setupFoo()' signature to
the outside, keep the creation of the guarded object were it is,
but allow it to be taken over bey a GuardedObject.

This setup still requires the _d_'tor to be public, but less
"dangerous" then a public c'tor - assuming there is some danger
in having that at all.

Use it for the PythoSettings and LanguageClientManager setup.

Change-Id: Ica114a0bb901d33d79c847910a2383ac1bc9e49a
Reviewed-by: David Schulz <david.schulz@qt.io>