Utils: Add overloads for CommandLine with osType
In the next commit we want to test these functions,
so we need to make sure that we can specify their
osType. We will also need these functions when we
are combining command lines from different platforms
e.g. on docker where the host might be windows,
but the docker image is Linux (or vice versa).

Task-number: QTCREATORBUG-32325
Change-Id: I7d01bdc5f292b6bf6b76d03856f6d4e588e371ab
Reviewed-by: hjk <hjk@qt.io>