ProjectExplorer: standard custom parser for build and run configurations
With this commit, a custom parser can be configured as the default
for build and run configurations of a project.

On the settings page for custom parser, additional columns are
displayed with which the relevant parser can be used as the
default for build and/or run configurations.

In the configurations, appropriately marked parsers are output
with the suffix “(project default)” and activated by default
for opened or newly created projects.

Fixes: QTCREATORBUG-32342
Change-Id: I2027fd3cc88cc4c38b08ac7ba6d921d950d448f9
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>