iOS: Fix potential crash with polling
There is the edge case when the user stops the application while the
polling devicectl call is running.
In that case, `DeviceCtlPollingRunner::stop` releases the poll task and
requests its deletion, but the same is done again in the `onDone`
implementation in `checkProcess`.

Add a check if the poll task still exists to the `onDone` handler in
`checkProcess`.

Change-Id: I930beeecbfcfdea04eba00d4e6a9c1f69766f5a1
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>