ProjectExplorer: Add convenience accessors to some classes
Namely, Kit, BuildConfiguration, BuildSystem and RunConfiguration get
functions to retrieve the active instance for a given project as well as
the one for the current and active projects, respectively.
This saves a ton of annoying null pointer checks.
Some of these functions replace existing ones with the same purpose that
were scattered around the code base.

Change-Id: I18247061229bbeb8b8ad1e15a2bd7a6626438bc0
Reviewed-by: hjk <hjk@qt.io>