iOS: Fix QML Profiling on devices with iOS <= 16
Re-implement IosDevice::portsGatheringRecipe(), which tried to run
netstat on the device. We don't know the used ports on device, the old
code used to just count up from 30000. Just do not report any used ports
for now.

The port on the device is mapped to a local port, to which the QML
Profiler actually has to connect. That is reported by iostool after
starting the app, so we need to adapt the runControl's qmlChannel
correspondingly when the correct port gets reported.

For the debug run worker we have to manually request the same logic for
QML debugging, with runControl->requestQmlChannel() if QML debugging is
enabled. We cannot use the base DebuggerRunTool::setupPortsGatherer(),
because that also requests that for the "debugChannel", and doing that
changes the whole logic how the DebuggerRunTool works, breaking
debugging in general on iOS.

Fixes: QTCREATORBUG-32403
Change-Id: I9574703e771e154fcfa7754706ac9871f3e61251
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>