iOS: Fix debugging and QML profiling on Simulator
Fixes a crash that was introduced by
a6aa050890789f689f3780cc5481f6e0340b5468 - the simulator device is not
derived from IosDevice, so we may not assume that.

Re-implement IosSimulator::portsGatheringRecipe - the base
implementation tries to run netstat "on" the simulator device, instead
do the same just locally, since the ports for the simulator are
effectively local ports. Also remove the custom port range for the same
reason.

Fixes: QTCREATORBUG-32416
Change-Id: I7294d0c20e94bafb9eb4590b7faaf6652792cbb1
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>