ClangTools: Create compilation databases per build configuration
The original idea was to have one per session and tool, but this runs
into issues due to the potentially different settings.
Even if that weren't the case, the "per build config" approach seems
saner and is in line with the general pattern that we want to have
everywhere in the future.

Fixes: QTCREATORBUG-32411
Change-Id: I8f59504099c40331698c61ea3e87a6abf832a670
Reviewed-by: David Schulz <david.schulz@qt.io>