Unarchiver: Improve ordering of methods
tar depends on other tools to extract anything else than tar itself, so
it is not reliable.
CMake does actually not rely on tar, but directly uses libarchive for
the various archives, so it is the more reliable choice, if available.
For example, tar on Linux might not be able to extract 7zip files, but
CMake can.

Change-Id: Ib5672557ada9619dd42697eb480eebc654a9d26d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>