lsp: do not restart servers shutting down
Do not restart lsp servers that crash during shutdown. That leads to
weird behaviors, for example a server that should be shutdown ends up
being restarted if it crashes during shutdown.

Context: qmlls from 6.8.1 returns 1 on successful exit, so Qt Creator
detects it as crash and restarts it (even when it was not supposed to be
running anymore). This behavior is fixed in qmlls from Qt 6.8.2.

Fixes: QTCREATORBUG-32428
Change-Id: I8a2449f8b395d8733a2c0faf826dd9b71e3b5dbd
Reviewed-by: David Schulz <david.schulz@qt.io>