QmlDesigner: Update the documentation for Run dropdown and Qt UI viewer
The Run dropdown button is a new inclusion which has
changed the way how applications can be run or preview
in Qt Design Studio. This document updates the basic
idea of that across a few documents. It also updates
the old Qt UI viewer document to include the new way
of intracting with the Android app and Qt design studio.

Fixes: QDS-13920
Fixes: QDS-14311
Fixes: QDS-13918
Change-Id: I66d3bdbf147c750f59d5521878015d6a8a7c0c92
Reviewed-by: Mats Honkamaa <mats.honkamaa@qt.io>