StringTable: Fix isQStringInUse()
With the current implementation it always returns true, even when
there is only one existing copy of a string inside m_strings set,
without any reference. This is due to creating a copy of the
QArrayDataPointer:

1 QArrayData::ref / qarraydata.h
2 QArrayDataPointer<char16_t>::ref / qarraydatapointer.h 438
3 QArrayDataPointer<char16_t>::QArrayDataPointer / qarraydatapointer.h 40
4 isQStringInUse

So, the ref_ is always at least 2.

Avoid creating a copy by taking a reference instead.

Observed misbehavior: After switching session nothing is erased from
the m_strings, it cumulates indefinitely instead.

Task-number: QTCREATORBUG-18800
Change-Id: I17981d44f88307e736ec03380baa39a03f0719bf
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>