QmlDesigner: fix IS_SUPPORTED_PROJECTSTORAGE_QT SCOPE
It can be a none cached variable in the PARENT_SCOPE
Depending code is reevaluated once the variable changed.
(Did not expect that, but that is one of the main magic of CMake)

Reduced FATAL_ERROR to a WARNING because PROJECTSTORAGE_QT
out of range Qt version users should still compile most of the code.

Change-Id: I83323775c8f69d8dd58a9a158797b280c8c99d68
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: Marco Bubke <marco.bubke@qt.io>