qmlls: only enable outline with experimental feature
Only enable the qmlls-powered outline when the
"Use advanced features (experimental)" checkbox is enabled. The new
outline is not feature complete compared to the previous outline (namely
the icons are different), so disable the outline feature by default.

Note that the outline feature only exists in qmlls from 6.9 and later.

Fixing the icons is tracked via QTCREATORBUG-31767.

Change-Id: I0f742ad8bacf026407e2ea7b151d04033ed273c6
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>