QmlDesigner: Remove qmltypes code path
With Qt 6.8 there we don't need the builtin.qmltypes anymore because
there is now the QML module. We used that code path for fake components,
too.

But that is problematic because we get errors if that components get
fixed in Qt.

Because we introduced already other ways, like unresolved type ids, to
handle that broken components, we should remove that code before we add
new features.

Task-number: QDS-14618
Change-Id: I82f3cdb77c44a0f1f8cf680763ddd1d4d7b9781c
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>