qmlls: make go-to-definition a default feature
go-to-definition is an "experimental" feature, but it is actually needed
to fix the bug at QTBUG-131920. Therefore, make it unexperimental and
use it by default, instead of using the buggy embedded codemodel.

I am not even completely sure why it was made experimental in the first
place, as its not even mentioned on the settings checkbox (so no string
changes needed there!).

[ChangeLog][plugins][qmljseditor] Qmlls will be used to process
go-to-definition requests without having to enable the
"Use advanced features" checkbox.

Task-number: QTBUG-131920
Change-Id: I1bb106022cc8d7611fe43fabbcd327d7d8573e62
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>