Python: Fix check for outdated pyls
Unfortunately pip still reports the python language server as outdated
even if the package was updated, most likely since the old package is
still in that directory. So instead of purely relying on the pip output
we compare the pip reported current version against the version
the python language server reports when startet with "--version". Since
neither the pip nor the version check is completely free we cache this
information now in case the server is installed or it is not and no pip
is available.

Change-Id: I3a56fd5240672555bd54e92685c0c58c4dc95d7c
Reviewed-by: Christian Stenger <christian.stenger@qt.io>