QmlDesigner: Project property sheet path collection
The property sheet paths are always collected if the designer directory,
module directory, or the qmldir file is changed. We could minimize the
updates, but then we must add an extra linking step in the project
storage. That extra linking step would be needed because there could be
property sheets for non-existing types.

Task-number: QDS-14618
Change-Id: Idc26e41e3b4298fb118452045f3d907ad9f5e619
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>