QtC: Refactor QmlJS::Icons
This is a preparation before using this refactored Icons::Provider
class in the QmlLSClient for providing the same icons in the
outline as it is provided currently by the QmlOutlineModel

The motivation behind this refactoring is the following:
1. Simplify memory management of Icons entity (using modern Meyers
 Singleton)
2. Encapsulate and group Loading(parsing) and querying of icons in
one entity (IconsStorage).
3. Make explicit that the loading of icons happens only once at the
 construction and stays immutable.
4. Extract free functions

Commenting a bit more on the 2nd,
Previously, the caller had to be aware of package names and the
internal icon storage structure, leading to hardcoded package names
and two queries at the caller site.
The current approach encapsulates this logic, bringing it closer to
parsing/loading, ensuring a single entity manages both folder
structure and query schema.
While this adds complexity (e.g., storing m_packageNames), it avoids
 duplicating knowledge.

Given that WebKit contains only one icon, a simpler <typeName, QIcon>
mapping could suffice.
However, for now, the more generalized approach is retained.

Task-number: QTCREATORBUG-31767
Change-Id: I073eaa82f1bc912ce7b40e3353aba23ddfcf8f4c
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>