Debugger: Fix dumper tests
`CheckSet` does not propagate its operators effects to the subchecks.
Making it work 100% correctly would require overriding all operators of
`Check` in `CheckSet`, which is a lot of code for few practical gains
(and also would make recently added `CrtpHelper` sort of useless).
All in all, I think now it is just easier to deal with the few cases
that we currently have at the individual check level rather than
trying to make it work for the checkset as a whole through
not-so-obvious CRTP tricks or tons of overrides and just wait for
more issues to accumulate before fixing them all for good.

Partially reverts 6ac1b0d00887a4bdce8beae796fbb986535f61c9.

Change-Id: Ibe2113db4cf69ffac8283c0e714603dc58ebb234
Reviewed-by: hjk <hjk@qt.io>