Fix crash when MinGW toolchains are detected on macOS
The "RealGcc" toolchain factory does all the autodetection, and happily
autodetects toolchains of various type, also MinGW toolchains on macOS
if installed e.g. via Homebrew. But, the MinGW toolchain factory is not
added on macOS. So we end up with toolchains that have a typeId that is
not supported by any toolchain factory, which leads to various asserts,
and finally to a crash.

The patch filters out any "GCC related" toolchains that actually do not
have a factory, during the autodetection in the "RealGCC" toolchain
factory. It also adds some QTC_ASSERTs, though these are in the end not
able to prevent crashes elsewhere (e.g. when opening the settings
dialog).

Fixes: QTCREATORBUG-32127
Change-Id: If5c17d1a3aff9df4ec4172c9ed7901990ac7f3ce
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>