QNX: Fix debugging
Before the amended change, when running QNX debugger, d->useDebugServer was false
(DebuggerRunTool::setUseDebugServer() wasn't called in this case),
so we called directly continueAfterDebugServerStart() from inside
startDebugServerIfNeededAndContinueStartup().

After the amended change, usedDebugChannel() returns true,
thus we are skipping direct execution of continueAfterDebugServerStart(),
and we are starting a process that never ends instead.

The fix is to introduce DebuggerRunParameters::m_skipDebugServer field,
defaulting to false, and to set it to true in case of qnx debugging.

Amends 1ebd3c1750a9044f753eaff5e343543c1983910f

Fixes: QTCREATORBUG-32478
Change-Id: I06c377ec6762371ab16582ae0fcdef6cacff1f83
Reviewed-by: hjk <hjk@qt.io>