TextEditor: Fix tab detection for short+simpler files
It was only detecting the usage of tabs if there were at least 3
*different* indentations with tabs. So a file that had all lines
starting with tab wasn't detected to have tab indentation.

In the case that we checked the whole file, it makes sense to assume
that any indentation that we find should be considered for the setting
for the whole file.

If we only took a sample of the whole file, we do not want to make a
guess for the whole file from only one or two indented lines.

Amends 08a66b778064ece5f01ee3323d8a5b5171b02c09

Change-Id: I615a35c7ad4e7e9ab3537eb76aad544603b47c4c
Reviewed-by: David Schulz <david.schulz@qt.io>