AutoTest: Allow debugging multiple tests
Remove the limitation of debugging to a single test case or function.
As when running tests, the configurations are created internally and
each test executable is started in debug mode sequentially.
Stopping the current debug session starts the next executable (if any),
stopping the whole test run stops the current debug and starts no
further
executables.

Change-Id: I9df5ae40f64361b808b92c3a572356213bd2851a
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>