EditorManager: Guard against potential crash
At this point `entry` should not be nullptr, since the editors passed to
closeEditor should be ones that are managed by EditorManager/
DocumentModel. But there could be misusage of the EditorManager API, or
some other corruption of the managed data in DocumentModel, so be
defensive.

Amends 119a0920c119e1bf631b16babab72b5c71740495

Task-number: QTCREATORBUG-32484
Change-Id: I3f173d1fe7c3533d4f044fd23171c69ab65bb9d8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>