CMakePM: PE: Hide (imported) suffix for buildtype names
Amends 67d22995968c99ab815d1499bc500c2f74187775

CMake Presets are implemented on top of build import functionality in Qt
Creator, since this can create kits / compilers etc.

But these build imports are artificial and should not be marked as
(imported) since this is something that Qt Creator does and not done by
the users.

Change-Id: I2ddf75d6fb3e4726f45b7431cacafc65f89bc39d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>