Docker: Move Container setup into its own thread
Since starting the container may happen from any thread, we need to
ensure that the container setup and teardown is done in the same thread.

This change moves the container setup into its own thread.

Without this, many asserts trigger as timers and other QObjects are
started and stopped from arbitrary threads.

Change-Id: I8db6f04358ee704ad4cd3e70dcc5bcfa7e237d27
Reviewed-by: hjk <hjk@qt.io>