designercore: optimize PlainTextEditModifier
There is no need to pass the QPlainTextEdit to the modifier. All
modifications can be done with the document. Also there is no need to
pass a text cursor as it is only used to join multiple edit blocks and
the position is completely unused we can construct a text cursor with
the passed document.

Change-Id: Iea726b3a01e7673be179398ab7a06058bd6a7392
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>