CMake: Use simpler and more correct code for running custom targets
Amends 5537a102e6df273afe63f9323410873831896c6d

The arguments for the CMake call that runs custom targets with the
`qtc_runnable` feature must be split with the rules for the platform
that they will be run on. The conceptually correct thing is to use the
OS type from the CMake executable (which is the build target info's
targetFilePath).

Using CommandLine with an empty executable is confusing and conceptually
wrong. Use ProcessArgs::joinArgs with the right OS type directly
instead.

Change-Id: Ibc0aa087d07922e8720b4304b9d3ba17d774cc7d
Reviewed-by: Cristian Adam <cristian.adam@qt.io>