Find: Make BaseTextFind more flexible
with regards to the underlying editor widget.
The TextEditor plugin wants to use its own custom text editor widget in
the future, so hardcoding the supported widget implementations prevents
the usage of BaseTextFind.

Make it a template class. Because templated QObject based classes cannot
have signals, basically rename the original implementation, add virtual
functions for the operations that are needed on the editor, and create a
new template subclass that relays these operations to the actual editor
type. (Similar to how we do that in other cases, like QFutureWatcher).

Change-Id: Idde1fb4dc19e755d379f8860d4b76f14410a18af
Reviewed-by: David Schulz <david.schulz@qt.io>