Core: Improve visibility of highlight marker for current view
In case of multiple editor views, the active view has a marker on top.
That marker was not very visible with the 2024 themes. Using
QPalette::Accent helps with it.

A hack restricts the usage of that color to those themes which also use
QPalette::Accent as highlight for the active mode in the mode bar. Other
themes continue to use QPalette::Highlight. Non-2024-themes are
therefore unaffected by this change.

Change-Id: I1e8c93243a6c69a4644726a7b1cdfaa1dc02b206
Reviewed-by: Eike Ziller <eike.ziller@qt.io>