Utils: Add contains with projection
Projections are a convienent way get a member variable or function of a
class. With that approach the value of the member can be tested directly
without a lambda etc..

struct Foo {
  QString message;

  QString title() const;
}

...

if (Utils::contains(foos, "hello"_L1, Foo::message))
 ...

or

if (Utils::contains(foos, "Re:"_L1, Foo::title))
 ...

The constraints prevent that there are collisions with other
implementations.

Change-Id: Ib26291d2cf852d56805cf1b851b835ef3edf4678
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>