ProjectExplorer: Better behavior for "go next/prev" in issues pane
... if no issues have associated files.
Normally, we only consider issues with associated files, as that allows
users to quickly navigate through the compiler errors' source locations.
However, if none of the issues have an associated file, that behavior
makes Creator appear to ignore the "next/prev" button. So we modify it
slightly so in such cases we consider all the entries.

Fixes: QTCREATORBUG-32503
Change-Id: I64f2331771e3f9a01e694bb7d6cfdc50332299f2
Reviewed-by: Christian Stenger <christian.stenger@qt.io>