ProjectExplorer: Differentiate between parsed and non-parsed tasks
... when linking from the issues pane to an output window.
If a task has corresponding content in e.g. the compile output pane and
it was *not* created by parsing that output (which is the case for qbs
error messages), then the output has not necessarily been written to the
text edit yet and we have to consider the queued output when determining
the line number.

Change-Id: I191b36d3ed913e1385e61e6cb1a223bb0ad30f87
Reviewed-by: Christian Stenger <christian.stenger@qt.io>