qmlls: don't check qmake versions for latest qmlls
Don't search for QMake versions in Qt kits when searching for the latest
qmlls version currently available.

Also clean up the code a bit, and break ties of Qt kits of the same
versions with their unique id, so that always the same qmlls is used
when multiple ones are available with the same version. This happens for
example when you have multiple Qt builds (one debug, one release, one
for documentation) based on the same source.

Change-Id: I0eb90d15c4e2b7c28c50f7ac1d3f976386b9b031
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>