AutoTest: Tweak result delegate
Treat secondary text differently except internal messages.
The first sentence is always made up by ourselves. Secondary text
is normally coming directly from the output the test produces.
Test frameworks usually format this text in a way that the output
will be aligned nicely on the console as it uses a monospaced font.

Fixes: QTCREATORBUG-32544
Change-Id: Ifd6029e6d44c9709650d3466bf64308f7062f296
Reviewed-by: David Schulz <david.schulz@qt.io>