Core: add an ui watchdog to QtSingleApplication
Adds the UI Watchdog of the KDToolBox to the QtSingleApplication that
periodically prints a debug message while the ui thread is locked with
the current duration of the lock if QTC_UI_WATCHDOG is defined in the
environment. Additionally this gives the oportunity to break whenever
the ui thread is locked.

Change-Id: I32aaa9f34822fa7f4a0dd1aa703fca49955b34a4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>