Support building with Sentry Native SDK
As a replacement for Crashpad directly. Enables us to report the version
number in a way that is useful in Sentry.

This might also open the option for informing the user with the option
to not send the report (with sentry_on_crash_handler, which on macOS is
only supported with Breakpad).

To use Sentry Native SDK set SENTRY_DSN to the API entry point,
SENTRY_PROJECT to the project in Sentry, and point the CMAKE_PREFIX_PATH
to the Sentry Native SDK installation path.

Both the Crashpad and the Breakpad backends are supported. Dependencies
(sentry dynamic library and crashpad_handler if used) are installed with
the `Dependencies` install component.

Change-Id: I3eb96cbee3ed3910b3f0dfe4c127bd1346b96fc6
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>