ExtensionSystem: Add Type="Script" to configure dependency resolution
We need some hint in the plugins metadata that tells us whether the
plugin is a script or a c++ plugin. Otherwise we cannot tell by the
metadata alone how to resolve dependencies. This is necessary for the
ExtensionBrowser which does not have any other way to detect the
type of an extension.

Change-Id: I61eb8f91fb079cf59cc8470f95a13c9c6fa75634
Reviewed-by: Eike Ziller <eike.ziller@qt.io>