Utils: Simplify ProcessArgs
Instead of trying to find a generic abstraction for the arguments
and splitting and combininig code paths several times, split into
the Windows and non-Windows case early once.

This also means that the ProcessArgs class as such is not really
used as value type anymore and could be converted to a namespace
in the future.

Change-Id: I79562b1571794d60cc97e5d7f877744f3bc7d57c
Reviewed-by: David Schulz <david.schulz@qt.io>