Git: Add gitignore on repo creation in existing project
When a git repository is created by the project wizard,
the .gitignore is created from a QtCreator template and
added to the "files to be committed" list.

Now perform the same when creating the repository later
in an already existing project directory.

If the project already contains a .gitignore, then use
this existing file and just mark it for committing to git.

Fixes: QTCREATORBUG-29776
Change-Id: Ie153c8dfb09a5640cf79941dfe7cfb608648a4b9
Reviewed-by: Orgad Shaneh <orgads@gmail.com>