Docker: Fix that GDB cannot write the index cache
The default setting for Docker devices is that we set the user ID for
operations on the device to the local host user ID. This makes sure that
when files are created by the Docker device in mounted local host
directories their user ID is sensible (and not root, which would be the
default).

Unfortunately that user ID (most probably) does not actually exist on
the Docker container, and the HOME directory defaults to the root
directory "/". GDB then tries to create its index cache directory in
"$HOME/.cache" which is not writable by the user ID.

Fix this by explicitly setting the HOME environment variable to
something writable when we use the local user ID on the Docker device.

Task-number: QTCREATORBUG-32562
Change-Id: Ice8be129c6b1da051438b49261ffcafec0864e04
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: hjk <hjk@qt.io>