Remove the Haskell plugin
Nowadays the same can be achieved by opening projects
as a Workspace. Users need to install and register the
language server themselves anyway and adding a build & run
configuration are trivial. So this plugin is a lot
of code that doesn't really improve the user experience over
the stock generic support.

Change-Id: Iad25a8e23b4fcbdaa5682f21bef6014156b07a5e
Reviewed-by: hjk <hjk@qt.io>