Docker: Wait until the container is actually started before returning
The call to "docker start" does not terminate, so this cannot be used to
wait until the container is actually started. It is also not enough to wait
until the "docker start" command is running, as this does not mean that the
container has actually been started. To wait until the container is
actually started we also start a "docker event" process that waits for a
started event of the container that is about to start. As this process will
only output a line when that container has started, we just have to wait
until the event listener process has some output and then terminate the
event listener.

Change-Id: I17008e8f5d19c2a647754a8ee31f1049811f0157
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>